Remove elastic/elastic-agent dependency
The main go.mod requires elastic/elastic-agent because
it is used by v2tool. This causes users of elastic-agent-client
to fetch the elastic-agent repo unnecessarily. This is doubly
bad because the version of elastic-agent required does not
exist; it probably came from a branch?

Move v2tool to its own module to prevent this dependency
leaking to users of elastic-agent-client. Update the elastic-agent
module version, and adjust the v2tool code accordingly.