Allow to set a port for the V2 stub server (#78)
Allow to set a port to `StubServerV2`. If its `Port` field is set,
this is the port used to open the TCP listener, if no port is set a
random one will be used. In all cases after `Start` is called, the
`Port` field will contain the port the server is listening on.