Add component-level configuration with max_procs limit
`CheckinExpected`/`CheckinObserved` now has the `component` field with
a component-level parameters. One of them is `limits` which contains the
`max_procs` that defines how many CPUs the component is allowed to
use.

This value is handled by the Go implementation of the client setting
`runtime.GOMAXPROCS` to `ComponentLimits.max_procs`.

The rest of the clients should implement their alternative to comply
the defined limit.