Add component-level configuration with go_max_procs limit (#83)
`CheckinExpected`/`CheckinObserved` now has the `component` field with a component-level parameters. One of them is `limits` which contains the `go_max_procs` that defines how many CPUs the component is allowed to use in a Go runtime.

This value is handled by the Go implementation of the client setting `runtime.GOMAXPROCS` to `ComponentLimits.go_max_procs`.

Non-Go clients should ignore this value.